docs: transfer the onclick block functions to a completeModel module
