+++ roulette/models/completeModel.js
@@ -1,4 +1,69 @@
-import './menuModel.js';
+import { minBet } from './consts/min_max.js';
+import { maxBet } from './consts/min_max.js';
+import { payoutRatios } from './consts/payoutRatios.js';
+import { rouletteNumber } from './consts/rouletteNumber.js';
+// import { rltPos } from './consts/rltPos.js';
+// import { rouletteSeries } from './consts/rouletteSeries.js';
+// import { oddDozColum } from './consts/oddDozColum.js';
+// import { fullBets } from './functions/fullBets.js';
 
+// для тестов:
 //export let strUp = 0;
 //export let completeBet = 25;
+
+// клик по кнопке "Рассчитать":
+calculate.onclick = function () {
+    let strUp = document.getElementById('str_up').value;
+    let completeBet = document.getElementById('complete').value;
+
+    // блок функций:
+    function chipsNeededForABet(strUp) {
+        let quantityNumb = payoutRatios.numb.position * rouletteNumber[strUp].numb;
+        let quantitySplit = payoutRatios.split.position * rouletteNumber[strUp].split;
+        let quantityCorner = payoutRatios.corner.position * rouletteNumber[strUp].corner;
+        let quantityStreet = payoutRatios.street.position * rouletteNumber[strUp].street;
+        let quantitySix_line = payoutRatios.six_line.position * rouletteNumber[strUp].six_line;
+        let quantitySum = quantityNumb + quantitySplit + quantityCorner + quantityStreet + quantitySix_line;
+        let result = quantitySum;
+        return result;
+    }
+    function sumBet(strUp) {
+        let sumChips = chipsNeededForABet(strUp);
+        let result = sumChips * completeBet;
+        return result;
+    };
+    function completePayment(strUp) {
+        let quantityNumb = payoutRatios.numb.position * rouletteNumber[strUp].numb * payoutRatios.numb.payout;
+        let quantitySplit = payoutRatios.split.position * rouletteNumber[strUp].split * payoutRatios.split.payout;
+        let quantityCorner = payoutRatios.corner.position * rouletteNumber[strUp].corner * payoutRatios.corner.payout;
+        let quantityStreet = payoutRatios.street.position * rouletteNumber[strUp].street * payoutRatios.street.payout;
+        let quantitySix_line = payoutRatios.six_line.position * rouletteNumber[strUp].six_line * payoutRatios.six_line.payout;
+        let quantitySum = quantityNumb + quantitySplit + quantityCorner + quantityStreet + quantitySix_line;
+        let result = quantitySum;
+        return result;
+    }
+    function totalPayment(strUp) {
+        let chips = completePayment(strUp);
+        let result = completeBet * chips;
+        return result;
+    };
+    function info(strUp) {
+        let color = rouletteNumber[strUp].color;
+        let parity = rouletteNumber[strUp].parity;
+        let magnitude = rouletteNumber[strUp].magnitude;
+        let dozen = rouletteNumber[strUp].dozen;
+        let column = rouletteNumber[strUp].column;
+        let result = `color: ${color} <br\/> \nparity: ${parity} <br\/> \nmagnitude: ${magnitude} <br\/> \ndozen: ${dozen} <br\/> \ncolumn: ${column}`;
+        return result;
+    };
+
+    // блок вызовов console.log:
+    console.log('strUp ' + strUp + ': ' + chipsNeededForABet(strUp) + ' positions of ' + completeBet);
+    console.log('sumBet: $' + sumBet(strUp));
+    console.log('payment: ' + completePayment(strUp) + ' chips');
+    console.log('totalPayment: $' + totalPayment(strUp));
+    console.log(info(strUp));
+
+    // вывод информации:
+    document.getElementById('info_1').innerHTML = `strUp  ${strUp}: ${chipsNeededForABet(strUp)} positions of ${completeBet}<br\/> sumBet: $${sumBet(strUp)}<br\/> payment: ${completePayment(strUp)} chips<br\/> totalPayment: $${totalPayment(strUp)}<br\/> ${info(strUp)}`;
+};
